Fix MSVC warnings in gnm and gcore
git-svn-id: https://svn.osgeo.org/gdal/trunk@31627 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/gcore/gdal_rat.cpp
@@ -1099,7 +1099,7 @@ void GDALDefaultRasterAttributeTable::AnalyseColumns()
 int GDALDefaultRasterAttributeTable::GetColumnCount() const
 
 {
-    return aoFields.size();
+    return static_cast<int>(aoFields.size());
 }
 
 /************************************************************************/
@@ -1840,7 +1840,7 @@ CPLErr GDALDefaultRasterAttributeTable::CreateColumn( const char *pszFieldName,
                                                GDALRATFieldUsage eFieldUsage )
 
 {
-    int iNewField = aoFields.size();
+    size_t iNewField = aoFields.size();
 
     aoFields.resize( iNewField+1 );
